requirements.txt
requests
google-genai
urllib3